% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-api.R
\name{api_params}
\alias{api_params}
\title{Request parameters}
\arguments{
\item{api_path}{The hypothesis API path, can be specify by `hypothesis.api.api_path` option or
`HYPOTHESIS_API_PATH` environment variable. Default: `https://hypothes.is/api/`.}

\item{endpoint}{API request endpoint}

\item{type}{API request type}

\item{api_key}{User api key, generated on the platform.}
}
\description{
Request parameters
}
