% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-api.R
\name{get_groups}
\alias{get_groups}
\title{Get a list of groups}
\source{
\url{https://h.readthedocs.io/en/latest/api-reference/v1/}
}
\usage{
get_groups(
  authority = NULL,
  document_uri = NULL,
  expand = NULL,
  api_path = get_hs_host(),
  api_key = NULL
)
}
\arguments{
\item{authority}{Filter returned groups to this authority. Default: main instance authority.}

\item{document_uri}{Document URI - only retrieve public (i.e. non-private) groups that apply
to a given document URI.}

\item{expand}{One or more relations to expand for a group resource Optional: `organization, scopes`}

\item{api_path}{The hypothesis API path, can be specify by `hypothesis.api.api_path` option or
`HYPOTHESIS_API_PATH` environment variable. Default: `https://hypothes.is/api/`.}

\item{api_key}{User api key, generated on the platform.}
}
\description{
Get a list of groups
}
