% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-api.R
\name{get_profile}
\alias{get_profile}
\title{Fetch profile information for the currently-authenticated user.}
\usage{
get_profile(api_path = get_hs_host(), api_key = NULL)
}
\arguments{
\item{api_path}{The hypothesis API path, can be specify by `hypothesis.api.api_path` option or
`HYPOTHESIS_API_PATH` environment variable. Default: `https://hypothes.is/api/`.}

\item{api_key}{User api key, generated on the platform.}
}
\description{
Fetch profile information for the currently-authenticated user.
}
