% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis-api.R
\name{get_profile_groups}
\alias{get_profile_groups}
\title{Fetch the groups for which the currently-authenticated user is a member.}
\usage{
get_profile_groups(api_path = get_hs_host(), api_key = NULL)
}
\arguments{
\item{api_path}{The hypothesis API path, can be specify by `hypothesis.api.api_path` option or
`HYPOTHESIS_API_PATH` environment variable. Default: `https://hypothes.is/api/`.}

\item{api_key}{User api key, generated on the platform.}
}
\description{
Fetch the groups for which the currently-authenticated user is a member.
}
