% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{hypothesis}
\alias{hypothesis}
\title{Hypothesis testing}
\usage{
hypothesis(df, y)
}
\arguments{
\item{df=}{dataframe}

\item{y=}{Binary target variable}
}
\description{
Takes dataframe having binary target variable
}
\author{
Sunny kumar and Vikas yadav
}
