% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{hypothesisBranding}
\alias{hypothesisBranding}
\title{Branding configuration}
\usage{
hypothesisBranding(
  accentColor,
  appBackgroundColor,
  ctaBackgroundColor,
  ctaTextColor,
  selectionFontFamily,
  annotationFontFamily
)
}
\arguments{
\item{accentColor}{Css-valid color value defining sidebar accent.}

\item{appBackgroundColor}{Css-valid color value defining sidebar background.}

\item{ctaBackgroundColor}{Css-valid color value defining main call-to-action button background.}

\item{ctaTextColor}{Css-valid color value defining font color used inside of the call-to-action buttons.}

\item{selectionFontFamily}{Css-valid color value defining font family used for annotation selection popup.}

\item{annotationFontFamily}{Css-valid color value defining font family used for annotation popup.}
}
\value{
Named list object storing provided arguments.
}
\description{
Configuration for custom annotations branding.
See more at: \href{https://h.readthedocs.io/projects/client/en/latest/publishers/config.html#cmdoption-arg-branding}{branding}.
}
