% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{hypothesisOnOff}
\alias{hypothesisOnOff}
\title{Turn On/Off Annotations Feature}
\usage{
hypothesisOnOff(
  initShow = TRUE,
  labelShow = "Enable Annotations",
  labelHide = "Disable Annotations"
)
}
\arguments{
\item{initShow}{Define initial state of annotations. When FALSE, annotations are turned off.}

\item{labelShow, labelHide}{Label to be shown in the button when annotations are hidden and shown respectively.}
}
\value{
A 'button' shiny.tag object, responsible for enabling/disabling of the package functionality.
}
\description{
Generates button that allows to turn on and turn off annotations tool.
}
