% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{hypothesisServices}
\alias{hypothesisServices}
\title{Hypothesis Service Configuration}
\usage{
hypothesisServices(
  apiUrl,
  authority,
  grantToken,
  allowLeavingGroups,
  enableShareLinks,
  groups,
  icon,
  onLoginRequest,
  onLogoutRequest,
  onSignupRequest,
  onProfileRequest,
  onHelpRequest
)
}
\arguments{
\item{apiUrl, authority, grantToken, allowLeavingGroups, enableShareLinks, groups, icon, onLoginRequest, onLogoutRequest, onSignupRequest, onProfileRequest, onHelpRequest}{Arguments defining external hypothesis services.}
}
\value{
Named list object storing provided arguments.
}
\description{
A list of alternative annotation services which the client should connect.
For more information visit: \href{https://h.readthedocs.io/projects/client/en/latest/publishers/config.html#cmdoption-arg-services}{services}.
}
