% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypo with binary.R
\name{hypothesis}
\alias{hypothesis}
\title{Hypothesis testing}
\usage{
hypothesis(x, y)
}
\arguments{
\item{x=}{dataframe}

\item{y=}{Binary target variable}
}
\value{
It will return DataFrame having P-value, t-statistic, X-statistic for all variable
}
\description{
Takes dataframe having binary target variable
}
\author{
Tejas Vartak, Shivani Goyal
}
